Move hooks
Moves hook from ui/components/hooks to ui/utils/hooks. Also makes sure that all of the import paths are fixed.

Resolves merge conflict with master: preserves useCallback-based ping stability fix from c1f9d39a6c9 while using @/components/graphql/queries/* path aliases for the new location. Also updates stale vi.mock() paths in connections/ConnectionTable.test.tsx and connections/metadata.test.tsx to the new @/utils/hooks/useKubernetesHook location.
