+++ ui/components/Dashboard/resources/crds/config.tsx
@@ -4,7 +4,7 @@ import { getK8sClusterIdsFromCtxId, getK8sContextFromClusterId } from '@/utils/m
 import { getAllCustomResourceDefinitionsKinds, SINGLE_VIEW } from '../config';
 import { Title } from '../../view';
 import { TooltipWrappedConnectionChip } from '../../../connections/ConnectionChip';
-import useKubernetesHook from '../../../hooks/useKubernetesHook';
+import useKubernetesHook from '@/utils/hooks/useKubernetesHook';
 import { DefaultTableCell, SortableTableCell } from '../sortable-table-cell';
 import { CONNECTION_KINDS } from '../../../../utils/Enum';
 import { FormatId } from '@/components/DataFormatter';
