+++ ui/components/Dashboard/resources/workloads/config.tsx
@@ -5,7 +5,7 @@ import { SINGLE_VIEW } from '../config';
 import { Title } from '../../view';
 import { TooltipWrappedConnectionChip } from '../../../connections/ConnectionChip';
 import { ResizableCell } from '../../../../utils/utils';
-import useKubernetesHook from '../../../hooks/useKubernetesHook';
+import useKubernetesHook from '@/utils/hooks/useKubernetesHook';
 import { DefaultTableCell, SortableTableCell } from '../sortable-table-cell';
 import { CONNECTION_KINDS } from '../../../../utils/Enum';
 import { FormatId } from '@/components/DataFormatter';
