+++ ui/components/MesherySettingsEnvButtons.tsx
@@ -19,9 +19,9 @@ import { useNotification } from '../utils/hooks/useNotification';
 import { EVENT_TYPES } from '../lib/event-types';
 import { CONNECTION_STATES } from '../utils/Enum';
 import { TooltipWrappedConnectionChip, ConnectionStateChip } from './connections/ConnectionChip';
-import useKubernetesHook from './hooks/useKubernetesHook';
+import useKubernetesHook from '@/utils/hooks/useKubernetesHook';
 import { keys } from '@/utils/permission_constants';
-import useTestIDsGenerator from './hooks/useTestIDs';
+import useTestIDsGenerator from '@/utils/hooks/useTestIDs';
 import CAN from '@/utils/can';
 import { useAddKubernetesConfigMutation } from '../rtk-query/connection';
 import { updateProgress } from '@/store/slices/mesheryUi';
