+++ ui/components/connections/metadata.tsx
@@ -14,7 +14,7 @@ import useKubernetesHook, {
   useMesheryOperator,
   useMeshsSyncController,
   useNatsController,
-} from '../hooks/useKubernetesHook';
+} from '@/utils/hooks/useKubernetesHook';
 import { TooltipWrappedConnectionChip } from './ConnectionChip';
 import { CONTROLLER_STATES, MESHSYNC_DEPLOYMENT_TYPE } from '../../utils/Enum';
 import { formatToTitleCase } from '../../utils/utils';
